pyproject.toml
[project]
name = "sneil-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]
